requirements.txt
mujoco==3.1.2
numpy==1.26.4
Pillow==10.2.0
transformers==4.38.2
torch==2.2.1
onnxruntime==1.17.1
openai==1.68.2
PyYAML==6.0.2
scipy==1.15.2
